+++ core/apis.go
@@ -115,6 +115,10 @@ func (s *administrationImpl) AdmGetAppOrgGroups(appID string, orgID string) ([]m
 	return s.app.admGetAppOrgGroups(appID, orgID)
 }
 
+func (s *administrationImpl) AdmGetAppOrgRoles(appID string, orgID string) ([]model.AppOrgRole, error) {
+	return s.app.AdmGetAppOrgRoles(appID, orgID)
+}
+
 func (s *administrationImpl) AdmGetApplicationPermissions(appID string, orgID string, l *logs.Log) ([]model.Permission, error) {
 	return s.app.admGetApplicationPermissions(appID, orgID, l)
 }